Bug #20971120 MTR TESTS FAIL WHILE CREATING SYSTEM TABLES ON NON PS BUILD
If compiling without performance schema, use a generated empty SQL file, that just contains an empty sys schema (like performance_schema does when not compiled in), to compile in to mysql_upgrade, mysql_install_db, and mysqld --initialize. Set the version as 1.0.0 for the sys schema, so that mysql_upgrade etc. will install the empty schema, but will not overwrite an existing schema.

mysql_upgrade and mysql_install_db still have the --skip-sys-schema options, and can still be used to not create the empty schema if desired.

Also add a --skip-sys-schema option to MTR to allow skipping use of the sys schema. This has to be done in this way, as we can't rely on checking the compile time option. It also leaves the possibility of not using the sys schema when performance schema is compiled in too, should people want to use it.

Like mysql_install_db etc., this currently creates an empty sys schema when used and removes the sysschema suite from the set of default suites to run.

Reviewed-by: Marc Alff <marc.alff@oracle.com>
Reviewed-by: Bjorn Munch <bjorn.munch@oracle.com>
RB: 9009