Fix for bug#21185585
SIGNALS DELIVERED TO INCORRECT OR GARBLED 'TRP_CLIENT' -> CLIENT CRASH OR HANGS

The 'TransporterFacade' in the API clients maintains a Vector
of valid client connections. As more API clients connects,
this Vector may be resized, which involves a reallocation
of the structure where the Clients are registered in.

There were no mechanisms to protect this reallocat from concurrent
'get' from the same Vector. Thus, garbled Client information
could be retrieved from this 'Vector of clients', causing
the signals to be delivered to some garbled memory structur believed
to represent client information.

This fix serialize expanding of the client Vector by moving
it into a function called when holding the poll right. 'get'ing
from the same Vector already does happen inside :trp_deliver_signal()
which holds the poll rights.