Dont use a removed bytes() method from page_size_t
The bytes() method was replaced with two methods - physical() and
logical() with the following meaning (from page0size.h):

        /* For non compressed tablespaces, physical page size is equal to
        the logical page size and the data is stored in buf_page_t::frame
        (and is also always equal to univ_page_size (--innodb-page-size=)).

        For compressed tablespaces, physical page size is the compressed
        page size as stored on disk and in buf_page_t::zip::data. The logical
        page size is the uncompressed page size in memory - the size of
        buf_page_t::frame (currently also always equal to univ_page_size
        (--innodb-page-size=)). */

So, the size of block->page.zip.data is page_size.physical() and the
size of block->frame is page_size.logical(), although in the case of non
compressed pages block->frame is what is in memory and what is stored on
disk and thus page_size.physical() == page_size.logical().

Thus, use page_size.physical() or page_size.logical() accordingly
instead of the nonexistent method page_size.bytes() which was removed.