WL#7816 - InnoDB: Persist the "corrupted" flag in the data dictionary
When InnoDB notices corruption in an index tree, it needs to flag it as
corrupted, both in its internal data dictionary cache and in persistent
storage.

With this change, the corruption bits are persisted by redo logs and
to a system table. Redo logs can be wrttien from low-level code and
background threads, such as page I/O handlers.

RB: 7497
Reviewed-by: Jimmy Yang <jimmy.yang@oracle.com>
Reviewed-by: Marko Makela <marko.makela@oracle.com>