BUG#21211524 CRASH IN ACTUAL_KEY_PARTS WITH PARTITIONED TABLES
Problem :
---------
In some rare cases optimizer would prune all partitions for a table
but won't remove the table and call storage apis to initialize and
fetch records. In this case we are not setting the active index during
initialization[ha_innopart::index_init].

For index scan with a begin key, we attempt to calculate key length
[Partition_helper::common_index_read] using active index which results
in the access violation. We return EOF error just after this in case
there is no partition to fetch data from[partition_scan_set_up].

Solution :
----------
Set active index id in ha_innopart::index_init even if there is
no partition to select from. This is also the way we handle in 5.6.

Reviewed-by: Mattias Jonsson <mattias.jonsson@oracle.com>

RB: 9231