Bug#21130285 MISSING MLOG_FILE_NAME RECORD FOR UNDO OR USER TABLESPACE
mtr_t::Command::prepare_write(): Invoke fil_names_write_if_was_clean()
on all tablespaces, even if a MLOG_FILE_NAME record was emitted for
the system or undo tablespace.

This is a regression from WL#7806, which is a follow-up to WL#7142.
In WL#7142, a mini-transaction commit could generate MLOG_FILE_NAME
for only one tablespace (the *.ibd file), while WL#7806 does it for
the system and undo tablespaces as well.

Reviewed-by: Jimmy Yang <jimmy.yang@oracle.com>