BUG#21111229 RESET SLAVE ALL BEHAVES DIFFERENT FOR DEFAULT AND              NON-DEFAULT CHANNELS BUG#21107331 RESET SLAVE ALL DOESN'T RESETS SLAVE_RETRIED_TRANSACTIONS
Since default channel should always exist, so its internal objects were not
deleted when executing RESET SLAVE ALL on default channel. But the internal
data was not re-initialized correctly, it caused these two bugs.

In this patch, we change the logic so that it deletes the original
internal objects(Master_info and Relay_log_info) of default channel
and then recreates them again when executing RESET SLAVE ALL on
default channel. In this way, it can guarantee all internal data is
re-initialized correctly.