Bug #20866059   CAN'T ENABLE EXPLICIT_DEFAULTS_FOR_TIMESTAMPS W ROLLING UPGRADES IN CHAINED RPL
The bug is essentially in that the binlog event applier (slave, or
mysqlbinlog client) miss out a necessary part of Query event execution
context. Specifically the part is

MASTER@@session.explicit_defaults_for_timestamp

value in the context of CREATE or ALTER table queries whose attribute
list includes at least one TIMESTAMP column.

***Here goes some analysis***

More specifically the TIMESTAMP attribute declaration that does not
have explicit NULL or NOT-NULL is subject to dual interpretation at
the query execution, which is controlable by the @@session var.
Which of the two methods (per the two var's values) was chosen by
the Master the binlog applier might have no idea.

The applier can learn about this type of execution context when events
are received (relay-logged) immediately from pre-WL6292 server thanks
to WL6292 introduced server version adaptation framework VAF (and
Bug19630322 fixing an MTS issue in it).

But once such events are relayed further downstream chain that info gets lost.

This patch addresses the following points of the issue.

 Legends:

 M - stands for Master server, S - for the applier (slave) server.
 The LaTeX style subscipt `_x' stands for a version `x'.
 `wl6292' is a version value
 corresponding to that of where the WL was pushed to.
 `fixed' is the current being fixed server version.
 `pre#' is for a version before that it prefixes.
 `[]' designates any in a range.
 `()' parameteries the server with a specific values of global var.

    Pre-WL6292 master events relayed downstream of chain replication:

M_pre#wl6292 -> S_fixed -> S_fixed -> ...

Data on S:s are consistent with the M.

    Cross "@@explicit_defaults_for_timestamp"-replication,
    from/to servers having different values of the var in the logger and in the applier.
    The chain replication is covered as well:

M_fixed(the-var := true)  -> S_fixed(false) -> S_fixed(true) ...

Notice that a problem of replicating from the WL6292 and later pre-fixed M
to fixed S is unsolvable in principle:

M_[wl6292..pre#fixed] --X-> S_fixed

because the var's value can't be learned by S_fixed even through VAF
(in contrast, on the pre#wl6292 servers the "var"'s value is false,
"" - as there's no such variable).
The patch treats this case with an assumption that M's value, awareness
of which is detected by VAF, is the same as the S' one.

In more details the first link of p.1 M_[0..pre#wl6292] -> S_fixed remains to be fixed by VAF
(see MTS fixes in bug19630322).

S_fixed -> S_fixed is fulfilled in the current bug's patch with
enriching the execution context that is recorded along with the Query
log event. The so called status var:s set is augmented with a specific
token when the master's parser faces a timestamp column
declaration. The token is essentially a boolean value
corresponding to @@session.explicit_defaults_for_timestamp.

Here is an example how the "old" intepretation settles down
timestamp column definition:

--server master
  start mysqld args: \empty

--connection master
select @@version;
=>  5.7.8-rc-debug-log
SELECT @@global.explicit_defaults_for_timestamp;
=> 0

CREATE TABLE tc (ts TIMESTAMP);
=> OK

show create table tc;
=> CREATE TABLE tc (
ts timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP
) ENGINE=InnoDB DEFAULT CHARSET=latin1

To recollect the docs, by rules of pre#WL6292 server which are in force in above
as @@global diagnoses, when the explicit NULL is not provided the
"implicit" NOT-NULL is added up to the actual column definition.
Moreover, the absence of the explicit NULL is the rules' necessary condition.

***Solution***

A solution that bases on exhausive analysis of parsed query details (presense of the explicit NULL)
would be perfect.
However that is rather tedious and the patch does not attempt precise investigation for that.
This could be always/possibly refined later.
A more coarse approach is taken to mark potentially vulnerable queries.

Dwelling into the low level details, a new three-enum-value member added to THD.
Two values correspond to of the two values of @@session.explicit_defaults_for_timestamp.
When the parser processes TIMESTAMP column declaration it memorizes the actual value of
the session var.
Whether the member's value is changed from the default UNSET is checked out
by the binary logger which further encodes it into a new Query-log-event status
variable of the bool sematic.
The meaning of the latter is to instruct the event applier about
the actual value of @@session.explicit_defaults_for_timestamp for executing
the Query-log-event.

Notice that @@explicit_defaults_for_timestamp is turned into writable
which is necessary because there's no other way to inform the "mysqlbinlog"
applier about the correct context but through the variable.

This type of approach is rather economic, extending only "suspicious"
Query-log-events with two bytes. A sign of the original context is
relayed along replication chain to lift out the reported issue in
particular.

The fixes enroll correction to main.mysqlbinlog test which dealt with
binlog files indexed explicitly. That turned to be not future-changes safe,
as the test limits max binlog size. And now when the new status var is added
the index have shifted.