BUG#15866285 FAILING ASSERTION: DICT_TF2_FLAG_IS_SET (INDEX->TABLE, DICT_TF2_TEMPORARY)
Problem:
INSERT causes assert in InnoDB as one of trigger successfully
changes the transaction mode to read only after a transaction
has already started.

Solution:
Make MySQL server adhere to following behavior, which is more
aligned with SQL standards and as per our current documentation.

- SET TRANSACTION always cause an exception if executed
inside a transaction, regardless of being submitted as a
standalone command, in procedures, functions or triggers.

- SET SESSION TRANSACTION always succeeds, and sets
characteristics which is applicable for subsequent transaction.

Test cases are added in main.commit.