Bug #20438524: 5.7.5 SKIP_NAME_RESOLVE STOPS ALL CONNECTION ATTEMPTS
In --skip-name-resolve move accounts with non-IP based host name parts
are in fact not available for one to login through.
The server generates a warning in the log file when it encounters one
such account and disables the account.

The checks to generate the warning for some reason were excluding the
"localhost" host name. Thus the default account created by e.g.
mysqld --initialize (root@localhost) was not triggering the warning
but people still couldn't login to it in --skip-name-resolve mode.

Fixed by:
1. Not excluding "localhost" when checking if a warning for the user
account should be written into the log.
2. Not disabling the accounts that people can't login against since
that also prevents proper use of SQL DEFINER stored programs that
reference these accounts.

Test case added.