Bug#21178589: REMOVE UNUSED FUNCTIONS AND CONVERT                GLOBAL SYMBOLS TO STATIC (LIBSQL)
The server defines some functions that are not called at all.
Other functions are called, but only from the same compilation unit.

Remove some function declarations and definitions, and add 'static'
keywords. Some symbols must be kept for separately compiled tools.

This is similar to Bug21141390, but for libsql rather than
libinnobase.