Bug#21251627 ASSERT `ORDER == __NULL && !SELECT_DISTINCT' AT JOIN::REPLACE_INDEX_SUBQUERY
We have IN(subquery on single table with group by and having).
remove_redundant_subquery_clauses() does not change this, because
of the presence of HAVING. If there had been ORDER BY it would
have removed it.
optimize_distinct_group_order() sees GROUP BY on primary key,
concludes that grouping is pointless, except that order still
has to be guaranteed (as GROUP BY is documented to produce ordered
groups), so it moves GROUP BY to ORDER BY. Now there is ORDER BY.
replace_index_subquery() sees IN(subquery) and dies here:
  // Guaranteed by remove_redundant_subquery_clauses():
  DBUG_ASSERT(order == NULL ...);
The flaw is that optimize_distinct_etc() adds a clause
after remove_etc().
Before the fix for Bug 21038929, optimize_distinct_etc() didn't add
the clause because it was a subquery; in general this is wrong reasoning,
but for IN(subquery) it's correct.
Fix: don't add the clause if IN(subquery).
Note that if we, in the future, stop guaranteeing that GROUP BY implies
ordering, then the modified code can be removed.