Bug#21097485: *insert_table_ref && (*insert_table_ref)->is_insertable
This problem occurs because we do not identify a single table
for insertion, which causes this assertion. The root cause is
that the column being assigned to is not checked for updatability.
Such check is therefore added when setting up the list of columns
to assign values to. (A column in an insertable-into view may still
be non-updatable).

This is achieved by centralizing the test for column updatability,
controlled by an extra argument to setup_fields(). (There are indeed
many calls that need updates, and if it wasn't for LOAD and one call
in UPDATE, we could have determined the property by looking at
required privileges, instead of adding this argument).

Bonus fix 1: A non-updatable column in INSERT does now cause the
correct error message "Column 'x' is not updatable" instead of
"The target table v of the INSERT is not insertable-into".

Bonus fix 2: All validation for non-updatable columns (except for
implicitly generated column lists and LOAD command) are now done in
setup_fields().

Bonus fix 3: Runtime checks for non-updatable columns were eliminated
from fill_record().

Bonus fix 4: Consistent error messages for non-updatable columns used
in LOAD command, and a test case. Error ER_LOAD_DATA_INVALID_COLUMN
is marked as unused.