Bug #21094069	FOREIGN KEYS WITH SPECIAL CHARS IN PARENT DB NAME: FALSE CONSTRAINT VIOLATIONTS *
PROBLEM

In function innobase_get_foreign_key_info(), we were trying to open the
referenced table using the name stored in key (Foreign_key) ,but the
table and database name strings are stored according the system
character set and not as file name charset format.Innodb stores the
names using filename charset format , therfore it is unable to
recognize the database and table names with special characters present
in the name and therefore is unable to open the referenced table and
we get subsequent error during inserts when foreign key checks is
turned on.

FIX

Convert the database and table names into filename charset and then
open the referenced table.