Backport of Part 2 (of 2) of fix for Bug#18390321 to 7.2 & 7.3
      MT-SCHEDULER: POSSIBLE BUSY WAIT IF PERFORMSEND() FAILED TO SEND

As 7.4 had some new features in this area of code, making
a plain merge impossible, a seperate patch has been made for 7.2 & 7.3.

It introduce the send-delay framework into 7.2. This already
existed in 7.4, but was refactored as part of the original patch.

As in the original patch, the overload delay has been
introduced into the plain delay-send framework.
See original push message for further comments.