BUG#20582189 - INNODB.DOUBLEWRITE FAILS DUE TO DATABASE CORRUPTION
This bug only happens sporadically on PB2, and mainly in Windows.
It's hard to reproduce. The test case writes some junk, such as 0,
to a data page to test if double write buffer works well. It looks like
the 'corrupted' page may not be flushed to disk immediately but buffered
in OS cache in Windows and the server would use a DIRECTIO to load page,
which makes the test case failure. So we just try to flush the data ASAP
in the test case, and try to print the page infomation on error.

RB: 9270
Reviewed-by: Jimmy Yang <jimmy.yang@oracle.com>