Bug#21212788 VARIABLE MAX_POINTS_IN_GEOMETRY TOO HIGH MAX VALUE
Issue: The max allowed number of points was too large and causes the server to
consume too much memory when computing a geometry buffer when given a valid
but large points per circle argument.

Fix: Reduce the MAX_POINTS_IN_GEOMETRY to 1024*1024 to limit the potential
memory and CPU consumption.