Combine inc() and dec() into a single method
The bodies of inc() and dec() were mostly the same, so combine them into
a common method and call it from inc() and dec().

Also use a busy loop trying to CAS the current value with the current
value + 1 (for inc() and -1 for dec()) instead of atomic increment (or
decrement) because that way we can check that we are not atomically
incrementing some of the magic values that could be used for m_val.
Currently it is only NOT_FOUND and this is not needed, but
GOTO_NEXT_ARRAY will be added soon.

WL#7170 InnoDB buffer estimates for tables and indexes