Fix for Bug#20957068:
NDB : EVENT API POLLEVENTS() MUST ACTUALLY CAUSE POLL OF TRANSPORTERS

The patch introduce the normal do_poll pattern to be used
by clients into ::pollEvent(). This also makes the condition member
in the EvenBuffers obsolete, as do_poll offers the exact same
functionality.

By doing its own poll of the receiver, we are not amu longer
dependent of other client threads or the ClusterMgr doing the
poll for us. This removes a problem where pollEvents were slow to
receive events.