Bug#17312761 MYSQLD WAITS FOR DUMMY NODES (NODEGROUP=65536) TILL NDB-WAIT-CONNECTED PERIOD
On startup mysqld waste time waiting on ndb data nodes that have
not joined the cluster.

Any ndbapi node including mysqld can only connect to a started
cluster.

Prior this patch mysqld (or a ndbapi application calling
wait_until_ready/2) waited until all configured data nodes were
started.

The ndbapi actually only need one connection to a data node with
DBTC to work, but this is not optimal since all data will be
routed internally in kernel via that data node.

Data nodes that not yet have joined the cluster will of course
never send any data until it have joined so no need to wait for
that node.

If a new data node eventually joins the cluster the ndbapi will
try to connect to the new node with an API-DB heartbeat interval.

With this patch mysqld will wait only for data nodes that have
joined the cluster.