BUG#18345363: ORDER OF APPLYING MYSQLBINLOG --DATABASE AND --REWRITE-DB IS WRONG BUG#20810442: "USE DATABSE_NAME" STATEMENTS OMITTED BY MYSQLBINLOG --REWRITE-DB OPTION
mysqlbinlog would apply the rewrite rules before applying the database
filter. This means that it in cases when mixing up SBR events with RBR
events then only one or the other would be output. This renders the
mysqlbinlog feature unusable in some use cases.

This patches changes the behavior so that the rewrite rules also apply
to the use clause, thence, making it possible to use the --database
on the query and row events. In addition, it removes the suppression
of the "use" clause.