Bug#21145277 VALGRIND ERRORS IN RQG_MDL_STRESS_VALGRIND AND RQG_MDL_DEADLOCK_VALGRIND
Valgrind reports errors while executing PFS_status_stats::aggregate_from(),
when under load, for an event scheduler thread.

The code should not even take this path,
as the event scheduler does not have THD associated with PFS_thread.

The root cause is in creat_thread(),
which does not properly initialize the PFS_instr::m_thd attribute.

As a result, a previous THD object from a previous thread,
which has been now freed, is used in the code,
leading to reads of free, invalid memory, and potentially to a crash.

The fix is to initialize PFS_instr::m_thd properly.