Go to the next array if the val is flagged as such
This patch handles GOTO_NEXT_ARRAY in the read scenarios (get(), del()).

This will be used when copying all the tuples to the next array.

Also, designate a deleted tuple with val == DELETED, instead of
key == DELETED. There are two reasons for this:
1. This way it is possible to atomically change a tuple from deleted to
   'go to the next array', by CASing val from DELETED to GOTO_NEXT_ARRAY.
2. This is more optimal in a scenario where a tuple with a given key is
   being constantly deleted and readded. Now a single cell in the array will
   be reused for this, instead of grabbing a new cell on each readd.

WL#7170 InnoDB buffer estimates for tables and indexes