Bug#20865683 IF AUTOCOMMIT=OFF AND GTID_NEXT='UUID:NO', GTID_MODE SWITCHING SHOULD BE BLOCKED
Problem:
When autocommit is set to off and gtid_next is set to 'UUID:NUMBER', then gtid_mode switch
was not being blocked.

Fix:
In addition to blocking a gtid_mode and an enforce_gtid_consistency switch when there is
an active multi-statement transaction, the switch is now also blocked when the thread owns
an anonymous transaction or the gtid_next is set to 'UUID:NUMBER'.
The check for 'gtid_next' was changed because: i) this situation is dealt with differently
when 'gtid_next' value is updated, and ii) the 'gtid_next' check only verifies if the thread
is in a multistament transaction if the transaction is not in the prepared state.