Bug #20581228: MYSQLD --HELP --VERBOSE TRIES TO LOCK FILES
Various issues are relevant to consider when running mysqld with
options --help --verbose:

1. On a loaded system, mysqld --help --verbose will initialize InnoDB in the
same way as for an ordinary start. It will try to lock the system files
present, and keeps trying for a hardcoded number of times, with a hardcoded
sleep inbetween. It will eventually fail, and this will also make mysqld
--help --verbose fail and report an error.

2. On an existing but empty datadir, InnoDB will create the system files
while initializing. These files will be left behind after mysqld --help
--verbose is finished, and will make a subsequent mysqld --initialize fail
due to the datadir not being empty.

3. On a non-existing datadir, the plugin table will not be found, and mysqld
--help --verbose will fail too.

InnoDB is initialized in order to enable reading the plugin table, which
contains the dynamic plugins. However, this is relevant only in issue 1
above. For issue 2 and 3, there is no point in initializing InnoDB since only
the builtin plugins will be listed anyway.

This patch applies the same constraint also in the first situation above.
Thus, we let mysqld --help --verbose only list options belonging to the
builtin plugins, and skip the dynamic ones. This allows us to execute
mysqld --help --verbose without initializing InnoDB and without opening the
plugin table, and this solves all three issues listed above.